+++ src/core/components/PostContent.tsx
