+++ src/core/partials/Certificate.tsx
