+++ src/core/partials/Footer.tsx
