+++ src/core/partials/PhotoViewer.tsx
