+++ src/layouts/Article.astro
@@ -19,3 +19,4 @@ const title = `${content.title} - ${AppConfig.site_name}`;
     <slot />
   </BlogPost>
 </Base>
+@/components/core/partials/BlogPost @/core
