+++ src/layouts/Base.astro
@@ -1,7 +1,7 @@
 ---
 import "../styles.css";
 
-import { Footer } from "@/components/partials/Footer";
+import { Footer } from "@/core/partials/Footer";
 import { AppConfig } from "@/utils/AppConfig";
 
 export interface Props {
@@ -33,4 +33,4 @@ const {
       <Footer />
     </div>
   </body>
-</html>
+</html>@/components/core/partials/Footer
