+++ src/pages/attendance-system.mdx
@@ -82,7 +82,7 @@ v2_images:
 preview_ids: ["images", "imagesv1"]
 ---
 
-import PhotoViewer from "@/components/partials/PhotoViewer";
+import PhotoViewer from "@/core/partials/PhotoViewer";
 
 ## Technology Overview
