+++ src/pages/drinkwell-apps.mdx
@@ -50,7 +50,7 @@ screnshots_2: ["1.png", "2.png", "3.png", "4.png", "5.png"]
 preview_ids: ["images", "images-2"]
 ---
 
-import PhotoViewer from "@/components/partials/PhotoViewer";
+import PhotoViewer from "@/core/partials/PhotoViewer";
 
 ## Technical Overview
