+++ src/pages/nmc-website.mdx
@@ -24,7 +24,7 @@ screnshots:
     alt: "Image Details"
 ---
 
-import PhotoViewer from "@/components/partials/PhotoViewer";
+import PhotoViewer from "@/core/partials/PhotoViewer";
 
 ## Technology Overview
