+++ src/pages/nybsys-website.mdx
@@ -32,7 +32,7 @@ screnshots:
     alt: "Image Details"
 ---
 
-import PhotoViewer from "@/components/partials/PhotoViewer";
+import PhotoViewer from "@/core/partials/PhotoViewer";
 
 ## Overview
